Former-commit-id: 06e2f33b2f6ea1535e003a5a0b2047a3dad3ad29

+++ src/gameengine/actors/properties/ActorDamageableWithExitProperty.java
@@ -15,7 +15,7 @@ public class ActorDamageableWithExitProperty<G extends ReadAndDamageGrid> extend
 	protected void damageActor(G grid, Integer actorID, Consumer<Double> damage, Double health) {
 		super.damageActor(grid, actorID, damage, health);
 		grid.getWriteableGameStatus().addExperience(health);
-		grid.getWriteableGameStatus().addMoney(health);
+		grid.getWriteableGameStatus().addMoney(health.intValue());
 	}
 
 }
